fix: satisfy pkg-file-count limits after layout persistence work
Consolidate layout validation/pruning sources, relocate layout and graph
topology tests into subpackages, and fold factory-dir resolution into
portable_bundled_files.go so package file-count checks pass.
