+++ pkg/factory/validation/profiletests/profile_test.go
@@ -1,4 +1,4 @@
-package validation_test
+package profiletests
 
 import (
 	"testing"
